+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Changed
 
+- concord-server: allow project-scoped secrets to be used when
+cloning the project's repositories;
 - concord-agent: improved error logging when cloning repositories;
 - concord-runner: fixed an issue preventing the runner from
 terminating correctly on `java.lang.Error`;